Update lokaltog fonts and X/dwm font handling
The lokaltog fonts have been renamed to lokaltog-fixed and
lokaltog-symbols. Dwm now uses lokaltog-fixed in the statusline and uses
icons stored in the Unicode Private Use area of lokaltog-symbols for the
statusline symbols.

The cureextra font has been removed because it's no longer in use.

All config files and scriptshave been updated to reflect these changes.